[refactor] inline_checkbox() を _common.py に移動
refs #51
